Revert "[test] Disable a couple of DebugInfo tests"
This reverts commit a182a7085de104bc19633d8479b9ce60d14abdb9.

I also fixed the tests as well.

NOTE: In the generic case of debug_value_addr, the self parameter with asserts
seems to be a var and without asserts a let. This is apparent at the AST level
before we even get to anything that I have changed. This seems to suggest that
there is some sort of uninitialized memory or something like that. Regardless
that is more of an AST level thing, so I just put in a regex that match both
patterns and send rdar://89237318 over to Robert Widmann!