[move-function] Propagate debug_value undef rather than treat it as an invalidation.
The reason that I am doing this is that otherwise in funclets we may show an
invalidated value being available if we have an llvm.dbg.addr for a reinit value
later in the funclet. This ensures that we have the undef at the beginning of
the funclet so before that llvm.dbg.addr we will properly show no value.