RequirementMachine: Allow mismatch in verify mode if there is a conflict
Having the Requirement Machine produce the same output as the
GenericSignatureBuilder when there are conflicting requirements
is difficult. It is also pointless, since invalid code does not
have a stable ABI.

Instead, just skip verification if the Requirement Machine
reported that there were conflicting rules in the current
minimization domain.