Diagnose 'case nil' in Non-Optional Switches
This is an anti-pattern since the resulting value will never compare equal to `nil`, and the entire switch-case is dead. This appears to be a misfeature as the subject value is simply type-checked against an optional which produces an injection which matches the global ~= for Optionals.

Effectively, `case nil:` becomes `case $match ~= nil`.

rdar://89742267