[cxx-interop] Test for apinotes fix in llvm-project.
This is the test that goes with https://github.com/apple/llvm-project/pull/4074

In the PR to apple/llvm-project we were bailing on DeclContexts that
were not isFileContext == true. This is problematic because with
C++-Interop enabled we can end up with a DeclContext that is an extern
"C". This test makes sure that the APINote is applied and that the code
from the 4074 PR no longer bails out.