RequirementMachine: Correctly handle importing of protocols in inferConditionalRequirements()
This fixes a regression from the rule sharing implementation; I forgot
to handle the imported rules in builder.ImportedRules here.

This makes the usage of RuleBuilder in conditional requirement
inference look like the other usages of RuleBuilder, except
that the results are passed to RewriteSystem::addRules() instead
of RewriteSystem::initialize().