a static let must be considered cross-actor
Static variables confer their isolation to their initializing expression.
At the same time, static variables are initialized on their dynamic first-use.

Thus, even if the static is a `let`, we have to assume all reads of the static let
might be the first one, and thus isolation must be acquired before reading the let.

This also means, for example, a non-sendable value can't be read from a static var
across actors, regardless of whether it is a let or not.

Resolves SR-16009 / rdar://problem/90436577