[AST] Move ModuleSearchPathLookup into ASTContext
`SearchPathOptions` should be a plain data object. Move `ModuleLookup`
into `ASTContext`, where we can then assume that the options and target
won't change from underneath us. This let's us remove all the handling
to check whether we need to rebuild the lookup table, now only building
it once when a lookup is requested (and adding to it if the search path
is modified after that point).

A further improvement could move `SearchPathsSet` into
`ModuleSearchPathLookup` and then remove the extra `SearchPathKind`.
Further still would be anything working with module paths, eg.
`getAllModuleSearchPathsSet` and the special cases in `ClangImporter`
that are mentioned in that method.