backtrack on part of SE-327 dealing with default-value exprs
This effectively reverts https://github.com/apple/swift/commit/68237447794d7a9aeeeed25422076387b84b3c4f

The blanket removal of isolation in default-value expressions had
unintented consequences for important workflows. It's still
a problem that needs to be addressed, but we need to be more precise
about the problematic situations.