RequirementMachine: Refactor shouldSplitConcreteEquivalenceClasses() and splitConcreteEquivalenceClasses() a bit
Note that this changes the behavior of a test slightly when the
-disable-concrete-contraction flag is used. This is because we're
not using the Requirement Machine that minimized the signature
and not the Requirement Machine built from the minimized
signature; the former includes a concrete conformance rule.

The isConcreteType() query returns true on the former when
given the generic parameter τ_0_0.

Since -disable-concrete-contraction is only meant for debugging,
I'm just removing that line from the test.