[swift-inspect] Add some basic tests for swift-inspect.
Swift-inspect requires root to be able to inspect other processes, which is not compatible with automated testing. To avoid this, we have the tests inspect the test process itself. We factor out most swift-inspect functionality into a static library, libswiftInspect. The executable and the tests both link this library, and the tests can then use it to inspect the test process, using the same code that the executable uses, and without requiring root.

Hook the tests into build-script by modifying build_script_helper.py to run swift-test when the testing is requested. The tests are written as normal tests in the package, and SwiftPM takes care of the rest.

The actual tests are very basic, only testing concurrency dumping. Concurrency dumping tests are limited to checking that no tasks or threads are discovered when no async code is running (actor detection currently has a lot of false positives, that's something we need to work on further), and that at least one task and thread is discovered when async code is running. Once we have a basic foundation set up we can start testing more functionality.

While we're at it, also improve Remote Mirror's "is actor" check by checking the context descriptor kind. This hasn't eliminated all false positives but it cut down on them a lot.