Fix handling of nested generic multi-payload enums
This particular case requires us to reprocess the metatype to "thicken" it.
That process inadvertently lost information about the depth of generic
nesting, which caused the RemoteMirror to be unable to resolve the type of
`C.E.t` at runtime.
```
  class C<T> {
    enum E<T> {
    case t(T)
    case u(Int)
    }
    var e: E<T>?
  }
```

This adds code to the thickening logic to preserve parent-type
information, which allows us to properly handle nested generic
types of this sort.

Resolves rdar://90490128