[NFC] Infrastructure to avoid storing simple generic params in metadata
The vast majority of generic parameters follow a simple pattern
(type parameter that defines a key argument), which we can take
advantage of to avoid actually representing them in type metadata:
if all the type arguments is the simple pattern, and there's not
more than some preset number, simply use a pre-allocated array in
the runtime.

Actually doing this for generic types will require some care around
back-deployment, as well as finding a bit to store that we've done
it.  But we can at least start doing this for new ABI, such as the
extended existential shape descriptors.