Add mayRetainUnmanagedReference deinit barrier
Adds new logic to deinit barrier to handle bitcasting to a reference.

Without this rule, all mutating methods on StringGuts are potentially
miscompiled because of the following property getters in StringObject:

    internal var discriminatedObjectRawBits: UInt64 {
      return Builtin.reinterpretCast(_object)
    }

    internal var largeAddressBits: UInt {
      return /* discriminatedObjectRawBits expression */
    }

    internal var nativeStorage: __StringStorage {
      return Builtin.reinterpretCast(largeAddressBits)
    }

Casting from reference to an integer and back inherently assumes that
the lifetime of 'self' is guaranteed for the duration of the
method. Hoisting destroy_addr in these methods miscompiles many tests
that rely on string. Hoisting destroy_addr is an optimization that has
been in tree for years across multiple releases. It was only an
arbitrary heuristics that kept it from miscompiling this case in the
past. Pure luck. This was recently exposed when destroy hoisting was
replaced with a version that didn't have the same heuristic.

Fixes rdar://90909833 (Add mayRetainUnmanagedReference deinit barrier)