[move-function] Update async dbginfo for upstream LLVM change.
I changed code upstream so that we are always guaranteed to have an entry value
for async arguments on r14. The main effect of this is that it eliminates a bunch of
cases where r14 would be spilled simplifying weird patterns in the test cases.

I also added additional commentary to the Dwarf part of the test cases
explaining what the patterns are checking and why.